showToast event, reload nav on change

+++ apps/kids/views.py
@@ -2,9 +2,11 @@ from datetime import date, datetime
 
 from django.http import HttpResponse
 from django.shortcuts import render
+from django.utils.translation import gettext_lazy as _
 from django.urls import reverse, reverse_lazy
 from django.views.decorators.http import require_http_methods
 from django.views.generic import CreateView, UpdateView
+from django_htmx.http import trigger_client_event
 from django_filters.views import FilterView
 from django_tables2 import SingleTableMixin
 
@@ -74,6 +76,14 @@ class ChildrenList(SingleTableMixin, FilterView):
         context["updated_obj"] = updated_obj
         return context
 
+    def get(self, request, *args, **kwargs):
+        response = super().get(request, *args, **kwargs)
+        created_obj = self.request.session.pop("created_obj", None)
+        if created_obj:
+            msg = _("Child has been added")
+            return trigger_client_event(response, "showToast", {"msg": msg})
+        return response
+
 
 class ChildCreateView(CreateView):
     model = Child
@@ -85,6 +95,10 @@ class ChildCreateView(CreateView):
         context = super().get_context_data(**kwargs)
         context["creating"] = True
         return context
+
+    def get_success_url(self):
+        self.request.session["created_obj"] = True
+        return super().get_success_url()
 
 
 class ChildUpdateView(UpdateView):
@@ -110,9 +124,12 @@ class ChildUpdateView(UpdateView):
 def delete_child(request, pk):
     if request.htmx:
         child_to_delete = Child.objects.filter(pk=pk).last()
-        parent = child_to_delete.parent
-        if parent.child_set.count() == 1:
-            parent.delete()
-        child_to_delete.delete()
-
-        return HttpResponse("")
+        if child_to_delete:
+            parent = child_to_delete.parent
+            if parent and parent.child_set.count() == 1:
+                parent.delete()
+            child_to_delete.delete()
+        resp = HttpResponse("")
+        msg = _("Child has been deleted.")
+        trigger_client_event(resp, "childDeleted")
+        return trigger_client_event(resp, "showToast", {"msg": msg})
